+++ CHANGELOG.md
@@ -4,14 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 - 2019-03-20
+
+### Added
+
+-   HubSpot Service
+-   Subscriptions can be saved as part of user save
+
 ## 0.0.1 - 2018-05-14
 
 ### Added
 
-*   Initial release
+-   Initial release
 
 ## 1.0.0 - 2018-05-17
 
 ### Added
 
-*   Documentation
+-   Documentation